refactor: use index exports for loadingmasks
